+++ static_precompiler/compilers/__init__.py
@@ -1,5 +1,5 @@
 # coding: utf-8
 from static_precompiler.compilers.coffeescript import CoffeeScript
-from static_precompiler.compilers.es6to5 import ES6To5
+from static_precompiler.compilers.babel import Babel
 from static_precompiler.compilers.scss import SASS, SCSS
 from static_precompiler.compilers.less import LESS