+++ static_precompiler/compilers/babel.py
@@ -3,15 +3,15 @@ from static_precompiler.compilers.base import BaseCompiler
 from static_precompiler.utils import run_command
 
 
-class ES6To5(BaseCompiler):
+class Babel(BaseCompiler):
 
-    name = "es6"
+    name = "babel"
     input_extension = "es6"
     output_extension = "js"
 
-    def __init__(self, executable="6to5"):
+    def __init__(self, executable="babel"):
         self.executable = executable
-        super(ES6To5, self).__init__()
+        super(Babel, self).__init__()
 
     def compile_file(self, source_path):
         return self.compile_source(self.get_source(source_path))