+++ static_precompiler/tests/test_babel.py
@@ -1,18 +1,18 @@
 # coding: utf-8
-from static_precompiler.compilers.es6to5 import ES6To5
+from static_precompiler.compilers.babel import Babel
 from static_precompiler.exceptions import StaticCompilationError
 from .test_coffeescript import clean_javascript
 import pytest
 
 
 def test_compile_file():
-    compiler = ES6To5()
+    compiler = Babel()
 
     assert clean_javascript(compiler.compile_file("scripts/test.es6")) == """"use strict";\nconsole.log("Hello, World!");"""
 
 
 def test_compile_source():
-    compiler = ES6To5()
+    compiler = Babel()
 
     assert clean_javascript(compiler.compile_source('console.log("Hello, World!");')) == """"use strict";\nconsole.log("Hello, World!");"""